+++ kamehouse-admin/src/main/java/com/nicobrest/kamehouse/admin/model/kamehousecommand/KeyPressKameHouseSystemCommand.java
@@ -1,7 +1,7 @@
 package com.nicobrest.kamehouse.admin.model.kamehousecommand;
 
 import com.nicobrest.kamehouse.commons.model.kamehousecommand.KameHouseSystemCommand;
-import com.nicobrest.kamehouse.commons.model.systemcommand.JvncSenderSystemCommand;
+import com.nicobrest.kamehouse.commons.model.systemcommand.TextJvncSenderSystemCommand;
 
 /**
  * KameHouseSystemCommand to send a key press.
@@ -14,7 +14,7 @@ public class KeyPressKameHouseSystemCommand extends KameHouseSystemCommand {
    * Sets the required SystemCommands to achieve this KameHouseSystemCommand.
    */
   public KeyPressKameHouseSystemCommand(KeyPress key, Integer keyPresses) {
-    String jvncSenderCommand = key.get(keyPresses);
-    systemCommands.add(new JvncSenderSystemCommand(jvncSenderCommand));
+    String text = key.get(keyPresses);
+    systemCommands.add(new TextJvncSenderSystemCommand(text));
   }
 }