+++ kamehouse-admin/src/main/java/com/nicobrest/kamehouse/admin/model/kamehousecommand/MouseClickKameHouseSystemCommand.java
@@ -2,7 +2,7 @@ package com.nicobrest.kamehouse.admin.model.kamehousecommand;
 
 import com.nicobrest.kamehouse.commons.exception.KameHouseInvalidCommandException;
 import com.nicobrest.kamehouse.commons.model.kamehousecommand.KameHouseSystemCommand;
-import com.nicobrest.kamehouse.commons.model.systemcommand.JvncSenderSystemCommand;
+import com.nicobrest.kamehouse.commons.model.systemcommand.MouseClickJvncSenderSystemCommand;
 
 /**
  * KameHouseSystemCommand to send a mouse click.
@@ -21,11 +21,20 @@ public class MouseClickKameHouseSystemCommand extends KameHouseSystemCommand {
    */
   public MouseClickKameHouseSystemCommand(Integer positionX, Integer positionY,
       Integer clickCount) {
+    this(positionX, positionY, clickCount, true);
+  }
+
+  /**
+   * Sets the required SystemCommands to achieve this KameHouseSystemCommand.
+   */
+  public MouseClickKameHouseSystemCommand(Integer positionX, Integer positionY,
+      Integer clickCount, boolean isLeftClick) {
     if (clickCount == null) {
       clickCount = 1;
     }
     validateParameters(positionX, positionY, clickCount);
-    systemCommands.add(new JvncSenderSystemCommand(positionX, positionY, clickCount));
+    systemCommands.add(
+        new MouseClickJvncSenderSystemCommand(positionX, positionY, clickCount, isLeftClick));
   }
 
   /**