+++ kamehouse-admin/src/main/java/com/nicobrest/kamehouse/admin/model/kamehousecommand/ScreenWakeUpKameHouseSystemCommand.java
@@ -1,10 +1,10 @@
 package com.nicobrest.kamehouse.admin.model.kamehousecommand;
 
 import com.nicobrest.kamehouse.commons.model.kamehousecommand.KameHouseSystemCommand;
-import com.nicobrest.kamehouse.commons.model.systemcommand.JvncSenderSystemCommand;
+import com.nicobrest.kamehouse.commons.model.systemcommand.MouseClickJvncSenderSystemCommand;
 
 /**
- * KameHouseSystemCommand to wake up the screen with a few mouse clicks.
+ * KameHouseSystemCommand to wake up the screen with a few left mouse clicks.
  *
  * @author nbrest
  */
@@ -14,8 +14,8 @@ public class ScreenWakeUpKameHouseSystemCommand extends KameHouseSystemCommand {
    * Sets the required SystemCommands to achieve this KameHouseSystemCommand.
    */
   public ScreenWakeUpKameHouseSystemCommand() {
-    systemCommands.add(new JvncSenderSystemCommand(400, 400, 1, 3));
-    systemCommands.add(new JvncSenderSystemCommand(400, 500, 1, 3));
-    systemCommands.add(new JvncSenderSystemCommand(500, 500, 1));
+    systemCommands.add(new MouseClickJvncSenderSystemCommand(400, 400, 1, 3));
+    systemCommands.add(new MouseClickJvncSenderSystemCommand(400, 500, 1, 3));
+    systemCommands.add(new MouseClickJvncSenderSystemCommand(500, 500, 1));
   }
 }