+++ kamehouse-admin/src/test/java/com/nicobrest/kamehouse/admin/integration/ScreenControllerIntegrationTest.java
@@ -35,7 +35,8 @@ class ScreenControllerIntegrationTest extends AbstractControllerIntegrationTest 
       "wakeUpTest, /wake-up",
       "unlockTest, /unlock",
       "keyPress, /key-press?key=ESC&keyPresses=1",
-      "mouseClick, /mouse-click?positionX=100&positionY=1000&clickCount=1"
+      "mouseClickLeft, /mouse-click?positionX=100&positionY=1000&clickCount=1&isLeftClick=true",
+      "mouseClickRight, /mouse-click?positionX=100&positionY=1000&clickCount=1&isLeftClick=false"
   })
   void screenControllerTest(String testName, String apiEndpoint) throws Exception {
     logger.info("Running {}", testName);