+++ kamehouse-cmd/src/main/java/com/nicobrest/kamehouse/cmd/executor/JvncSenderExecutor.java
@@ -32,7 +32,12 @@ public class JvncSenderExecutor implements Executor {
         int positionX = Integer.parseInt(mouseClickParams[0]);
         int positionY = Integer.parseInt(mouseClickParams[1]);
         int clickCount = Integer.parseInt(mouseClickParams[2]);
-        vncSender.sendMouseClick(positionX, positionY, clickCount);
+        boolean isLeftClick = Boolean.valueOf(mouseClickParams[3]);
+        if (isLeftClick) {
+          vncSender.sendMouseLeftClick(positionX, positionY, clickCount);
+        } else {
+          vncSender.sendMouseRightClick(positionX, positionY, clickCount);
+        }
       } else {
         vncSender.sendText(text);
       }